api/requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==2.2.6
requests>=2.28.1
matplotlib
flask
Pillow==12.0.0